Current

+++ bin/set_service_provider.sh
@@ -10,7 +10,10 @@
  #then
  #	rm /opt/engines/run/services/$1
  #fi
- rm -r /opt/engines/run/services/$1
+ if test -d /opt/engines/run/services/$1
+  then
+ 	rm -r /opt/engines/run/services/$1
+ fi
 cp -r /opt/engines/run/services-available/$1/$2 /opt/engines/run/services/$1
 chgrp -R containers /opt/engines/run/services/$1
 chmod g+rwx /opt/engines/run/services/$1
